change tests: export Heap
