+++ src/Heap.es6
@@ -1,18 +1,8 @@
 
 import 'core-js/shim';
 
-export default class Person {
+export default class Heap {
 
-  constructor( name ) {
-    this.name = name;
-  }
-
-  sayHello() {
-    return `Hello ${ this.name }!`;
-  }
-
-  sayHelloThreeTimes() {
-    let hello = this.sayHello();
-    return `${ hello } `.repeat(3);
+  constructor(options) {
   }
 }
